+++ src/onehealth/dhis2/__init__.py
@@ -1,0 +1,7 @@
+"""DHIS2 Web API integration."""
+
+from onehealth.dhis2.client import DHIS2APIError, DHIS2Client
+from onehealth.dhis2.mapping import DHIS2Mapping
+
+__all__ = ["DHIS2APIError", "DHIS2Client", "DHIS2Mapping"]
+